Add files via upload
The program will create a Java Swing application that displays a window with a Text field and three buttons: "Up", "Down" and Reset. When the "Up" button is clicked, the label will display the next number in the sequence 1, 2, 3,... up to 7,0,1 and so on, and when the "Down" button is clicked, the label should display the previous number in the sequence(upto 0, then again back to 7,6,5 and so on). When Reset button is clicked the label should be display 0.